requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.9.1
click==8.1.7
Django==4.2.5
django-debug-toolbar==4.2.0
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
Markdown==3.4.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.3
sqlparse==0.4.4
uritemplate==4.1.1
